Abort command execution when a command fails by default
It's pretty meaningless to keep executing commands when one command fails. So
let's make it abort by default if any command returns a non-zero exit code.

E.g.:

    $ tox
    py27 runtests: PYTHONHASHSEED='2154811636'
    py27 runtests: commands[0] | echo 1
    1
    py27 runtests: commands[1] | false
    ERROR: InvocationError: '/usr/bin/false'
    ERROR: Stopping processing of commands for env py27 because `/usr/bin/false` failed with exit code 1
    _____________________________________________ summary ______________________________________________
    ERROR:   py27: commands failed

Unless the `ignore_errors` setting is set to `True`, in which case, it keeps
processing, like it did in older versions of tox.